Refactored database related classes
